+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Fix
 
 - Fixed a non-critical bug that resulted in less-than optimal layouting when UTF-8 characters were involved.
+- No longer panic when rendering a `Table` that contains a row with `Row::max_height(0)`.
 
 ## [7.2.2] - 2026-01-13